feat: Fork plugin to G4EUkrChatSupportFork
Complete project rename to avoid conflict with the official Dalamud repository. Renamed all files, folders, namespaces, and updated project/solution files and the custom repository manifest.
