+++ G4EUkrChatSupportFork/G4EUkrChatSupportFork.json
@@ -1,9 +1,9 @@
 {
     "Author": "evercreepy",
-    "Name": "G4E UkrChatSupport",
+    "Name": "G4E UkrChatSupport (Fork)",
     "Punchline": "Support for ukrainian characters.",
     "Description": " Original plugin by justscribe. Support for ukrainian characters in text chat and input field.",
-    "InternalName": "G4EUkrChatSupport",
+    "InternalName": "G4EUkrChatSupportFork",
     "ApplicableVersion": "any",
     "DalamudApiLevel": 12,
     "RepoUrl": "https://github.com/evercreepy/xiv_ukrchatsupport",
