+++ repo.json
@@ -1,15 +1,15 @@
 [
     {
         "Author": "justscribe",
-        "Name": "G4EUkrChatSupport",
+        "Name": "G4EUkrChatSupportFork",
         "Description": "Support for ukrainian characters in text chat and input field.",
-        "InternalName": "G4EUkrChatSupport",
+        "InternalName": "G4EUkrChatSupportFork",
         "AssemblyVersion": "1.0.0.35",
         "TestingAssemblyVersion": "1.0.0.35",
         "RepoUrl": "https://github.com/evercreepy/xiv_ukrchatsupport",
         "ApplicableVersion": "any",
-        "IsHide": "False",
-        "IsTestingExclusive": "False",
+        "IsHide": false,
+        "IsTestingExclusive": false,
         "DownloadCount": 0,
         "LastUpdate": 0,
         "LoadPriority": 69420,
